move serializers to a serializers package
