+++ blueflood-core/src/main/java/com/rackspacecloud/blueflood/io/serializers/StringMetadataSerializer.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package com.rackspacecloud.blueflood.io;
+package com.rackspacecloud.blueflood.io.serializers;
 
 import com.google.protobuf.CodedInputStream;
 import com.google.protobuf.CodedOutputStream;
