+++ blueflood-core/src/test/java/com/rackspacecloud/blueflood/io/CounterRollupSerializationTest.java
@@ -16,6 +16,7 @@
 
 package com.rackspacecloud.blueflood.io;
 
+import com.rackspacecloud.blueflood.io.serializers.NumericSerializer;
 import com.rackspacecloud.blueflood.types.CounterRollup;
 import junit.framework.Assert;
 import org.apache.commons.codec.binary.Base64;
